refactor: rush 일반 타입 - api 타입 분리
